# Tekst-API/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tekst-Dev",
    "description": "An online text research platform",
    "termsOfService": "None",
    "contact": {},
    "license": {
      "name": "AGPL-3.0-or-later",
      "url": "https://www.gnu.org/licenses/agpl-3.0.txt"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1/api"
    }
  ],
  "paths": {
...
    "/browse/content-siblings": {
      "get": {
        "tags": [
          "browse"
        ],
        "summary": "Get content siblings",
        "description": "Returns a list of all resource contents belonging to the resource\nwith the given ID, associated to locations that are children of the parent location\nwith the given ID.\n\nAs the resulting list may contain contents of arbitrary type, the\nreturned content objects cannot be typed to their precise resource content type.",
        "operationId": "getContentSiblings",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "res",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "5eb7cf5a86d9755df3a6c593",
              "description": "ID of resource the requested contents belong to",
              "title": "Res"
            },
            "description": "ID of resource the requested contents belong to"
          },
          {
            "name": "parent",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "example": "5eb7cf5a86d9755df3a6c593"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ID of location for which siblings to get contents for",
              "title": "Parent"
            },
            "description": "ID of location for which siblings to get contents for"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/DebugContentRead"
                      },
                      {
                        "$ref": "#/components/schemas/PlaintextContentRead"
                      }
                    ],
                    "discriminator": {
                      "propertyName": "resourceType",
                      "mapping": {
                        "debug": "#/components/schemas/DebugContentRead",
                        "plaintext": "#/components/schemas/PlaintextContentRead"
                      }
                    }
                  },
                  "title": "Response Get Content Siblings Browse Content Siblings Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/browse/location-data": {
      "get": {
        "tags": [
          "browse"
        ],
        "summary": "Get location data",
        "description": "Returns the location path from the location with the given level/position\nas the last element, up to its most distant ancestor location\non structure level 0 as the first element of an array as well as all contents\nfor the given resource(s) referencing the locations in the location path.",
        "operationId": "getLocationData",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "txt",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "example": "5eb7cf5a86d9755df3a6c593",
              "description": "ID of text to look up data for",
              "title": "Txt"
            },
            "description": "ID of text to look up data for"
          },
          {
            "name": "lvl",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Location level",
              "title": "Lvl"
            },
            "description": "Location level"
          },
          {
            "name": "pos",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Location position",
              "title": "Pos"
            },
            "description": "Location position"
          },
          {
            "name": "res",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "example": "5eb7cf5a86d9755df3a6c593"
              },
              "description": "ID (or list of IDs) of resource(s) to return content data for",
              "default": [],
              "title": "Res"
            },
            "description": "ID (or list of IDs) of resource(s) to return content data for"
          },
          {
            "name": "head",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "Only return contents for the head location of the path",
              "default": false,
              "title": "Head"
            },
            "description": "Only return contents for the head location of the path"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Return at most <limit> contents",
              "default": 4096,
              "title": "Limit"
            },
            "description": "Return at most <limit> contents"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocationData"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/browse/nearest-content": {
      "get": {
        "tags": [
          "browse"
        ],
        "summary": "Get nearest content",
        "description": "Finds the nearest location the given resource holds content for and returns\ndata for that location.",
        "operationId": "getNearestContent",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "pos",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Location position",
              "title": "Pos"
            },
            "description": "Location position"
          },
          {
            "name": "targetRes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "example": "5eb7cf5a86d9755df3a6c593",
              "description": "ID of resource to return nearest location with content for",
              "default": [],
              "title": "Targetres"
            },
            "description": "ID of resource to return nearest location with content for"
          },
          {
            "name": "res",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string",
                "example": "5eb7cf5a86d9755df3a6c593"
              },
              "description": "ID (or list of IDs) of resource(s) to return content data for",
              "default": [],
              "title": "Res"
            },
            "description": "ID (or list of IDs) of resource(s) to return content data for"
          },
          {
            "name": "mode",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "preceding",
                "subsequent"
              ],
              "type": "string",
              "description": "Whether to look for the nearest preceding or subsequent location with content",
              "default": "subsequent",
              "title": "Mode"
            },
            "description": "Whether to look for the nearest preceding or subsequent location with content"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Return at most <limit> contents",
              "default": 4096,
              "title": "Limit"
            },
            "description": "Return at most <limit> contents"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "title": "Response Get Nearest Content Browse Nearest Content Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/browse/locations/{id}/path/options-by-head": {
      "get": {
        "tags": [
          "browse"
        ],
        "summary": "Get path options by head id",
        "description": "Returns the options for selecting text locations derived from the location path of\nthe location with the given ID as head.",
        "operationId": "getPathOptionsByHeadId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "5eb7cf5a86d9755df3a6c593",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/LocationRead"
                    }
                  },
                  "title": "Response Get Path Options By Head Id Browse Locations  Id  Path Options By Head Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/browse/locations/{id}/path/options-by-root": {
      "get": {
        "tags": [
          "browse"
        ],
        "summary": "Get path options by root",
        "description": "Returns the options for selecting text locations derived from the location path of\nthe location with the given ID as root. At each level, the first option is taken\nas the basis for the next level.",
        "operationId": "getPathOptionsByRoot",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "5eb7cf5a86d9755df3a6c593",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/LocationRead"
                    }
                  },
                  "title": "Response Get Path Options By Root Browse Locations  Id  Path Options By Root Get"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/browse/resources/{id}/coverage": {
      "get": {
        "tags": [
          "browse"
        ],
        "summary": "Get resource coverage data",
        "operationId": "getResourceCoverageData",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "5eb7cf5a86d9755df3a6c593",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceCoverage"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/browse/resources/{id}/coverage-details": {
      "get": {
        "tags": [
          "browse"
        ],
        "summary": "Get detailed resource coverage data",
        "operationId": "getDetailedResourceCoverageData",
        "security": [
          {
            "APIKeyCookie": []
          },
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "5eb7cf5a86d9755df3a6c593",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResourceCoverageDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DebugContentRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "example": "5eb7cf5a86d9755df3a6c593"
          },
          "resourceId": {
            "type": "string",
            "title": "Resourceid",
            "description": "Resource ID",
            "example": "5eb7cf5a86d9755df3a6c593"
          },
          "resourceType": {
            "const": "debug",
            "title": "Resourcetype"
          },
          "locationId": {
            "type": "string",
            "title": "Locationid",
            "description": "Parent text location ID",
            "example": "5eb7cf5a86d9755df3a6c593"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment",
            "description": "Plaintext, potentially multiline comment on this content"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text",
            "description": "Text content of the debug content"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "resourceId",
          "resourceType",
          "locationId"
        ],
        "title": "DebugContentRead"
      },
...
      "PlaintextContentRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "example": "5eb7cf5a86d9755df3a6c593"
          },
          "resourceId": {
            "type": "string",
            "title": "Resourceid",
            "description": "Resource ID",
            "example": "5eb7cf5a86d9755df3a6c593"
          },
          "resourceType": {
            "const": "plaintext",
            "title": "Resourcetype"
          },
          "locationId": {
            "type": "string",
            "title": "Locationid",
            "description": "Parent text location ID",
            "example": "5eb7cf5a86d9755df3a6c593"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment",
            "description": "Plaintext, potentially multiline comment on this content"
          },
          "text": {
            "type": "string",
            "maxLength": 102400,
            "minLength": 1,
            "title": "Text",
            "description": "Text content of the plaintext content"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "resourceId",
          "resourceType",
          "locationId",
          "text"
        ],
        "title": "PlaintextContentRead"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tekstuserauth"
      },
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/jwt/login"
          }
        }
      }
    }
  },
...
}